LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/animations/font-style-interpolation-expected.txt

TT

FAIL CSS Transitions: property <font-style> from [normal] to [oblique 10deg] at (-2) should be [oblique -20deg] assert_equals: expected "oblique - 20deg " but got "oblique 10deg "
FAIL CSS Transitions: property <font-style> from [normal] to [oblique 10deg] at (-0.25) should be [oblique -2.5deg] assert_equals: expected "oblique - 2.5deg " but got "oblique 10deg "
FAIL CSS Transitions: property <font-style> from [normal] to [oblique 10deg] at (0) should be [normal] assert_equals: expected "normal " but got "oblique 10deg "
FAIL CSS Transitions: property <font-style> from [normal] to [oblique 10deg] at (0.3) should be [oblique 3deg] assert_equals: expected "oblique 3deg " but got "oblique 10deg "
FAIL CSS Transitions: property <font-style> from [normal] to [oblique 10deg] at (0.6) should be [oblique 6deg] assert_equals: expected "oblique 6deg " but got "oblique 10deg "
PASS CSS Transitions: property <font-style> from [normal] to [oblique 10deg] at (1) should be [oblique 10deg]
FAIL CSS Transitions: property <font-style> from [normal] to [oblique 10deg] at (1.5) should be [oblique 15deg] assert_equals: expected "oblique 15deg " but got "oblique 10deg "
FAIL CSS Transitions with transition: all: property <font-style> from [normal] to [oblique 10deg] at (-2) should be [oblique -20deg] assert_equals: expected "oblique - 20deg " but got "oblique 10deg "
FAIL CSS Transitions with transition: all: property <font-style> from [normal] to [oblique 10deg] at (-0.25) should be [oblique -2.5deg] assert_equals: expected "oblique - 2.5deg " but got "oblique 10deg "
FAIL CSS Transitions with transition: all: property <font-style> from [normal] to [oblique 10deg] at (0) should be [normal] assert_equals: expected "normal " but got "oblique 10deg "
FAIL CSS Transitions with transition: all: property <font-style> from [normal] to [oblique 10deg] at (0.3) should be [oblique 3deg] assert_equals: expected "oblique 3deg " but got "oblique 10deg "
FAIL CSS Transitions with transition: all: property <font-style> from [normal] to [oblique 10deg] at (0.6) should be [oblique 6deg] assert_equals: expected "oblique 6deg " but got "oblique 10deg "
PASS CSS Transitions with transition: all: property <font-style> from [normal] to [oblique 10deg] at (1) should be [oblique 10deg]
FAIL CSS Transitions with transition: all: property <font-style> from [normal] to [oblique 10deg] at (1.5) should be [oblique 15deg] assert_equals: expected "oblique 15deg " but got "oblique 10deg "
FAIL CSS Animations: property <font-style> from [normal] to [oblique 10deg] at (-2) should be [oblique -20deg] assert_equals: expected "oblique - 20deg " but got "normal "
FAIL CSS Animations: property <font-style> from [normal] to [oblique 10deg] at (-0.25) should be [oblique -2.5deg] assert_equals: expected "oblique - 2.5deg " but got "normal "
PASS CSS Animations: property <font-style> from [normal] to [oblique 10deg] at (0) should be [normal]
FAIL CSS Animations: property <font-style> from [normal] to [oblique 10deg] at (0.3) should be [oblique 3deg] assert_equals: expected "oblique 3deg " but got "normal "
FAIL CSS Animations: property <font-style> from [normal] to [oblique 10deg] at (0.6) should be [oblique 6deg] assert_equals: expected "oblique 6deg " but got "oblique 10deg "
PASS CSS Animations: property <font-style> from [normal] to [oblique 10deg] at (1) should be [oblique 10deg]
FAIL CSS Animations: property <font-style> from [normal] to [oblique 10deg] at (1.5) should be [oblique 15deg] assert_equals: expected "oblique 15deg " but got "oblique 10deg "
FAIL Web Animations: property <font-style> from [normal] to [oblique 10deg] at (-2) should be [oblique -20deg] assert_equals: expected "oblique - 20deg " but got "normal "
FAIL Web Animations: property <font-style> from [normal] to [oblique 10deg] at (-0.25) should be [oblique -2.5deg] assert_equals: expected "oblique - 2.5deg " but got "normal "
PASS Web Animations: property <font-style> from [normal] to [oblique 10deg] at (0) should be [normal]
FAIL Web Animations: property <font-style> from [normal] to [oblique 10deg] at (0.3) should be [oblique 3deg] assert_equals: expected "oblique 3deg " but got "normal "
FAIL Web Animations: property <font-style> from [normal] to [oblique 10deg] at (0.6) should be [oblique 6deg] assert_equals: expected "oblique 6deg " but got "oblique 10deg "
PASS Web Animations: property <font-style> from [normal] to [oblique 10deg] at (1) should be [oblique 10deg]
FAIL Web Animations: property <font-style> from [normal] to [oblique 10deg] at (1.5) should be [oblique 15deg] assert_equals: expected "oblique 15deg " but got "oblique 10deg "
FAIL CSS Transitions: property <font-style> from [oblique 5deg] to [oblique 15deg] at (-2) should be [oblique -15deg] assert_equals: expected "oblique - 15deg " but got "normal "
PASS CSS Transitions: property <font-style> from [oblique 5deg] to [oblique 15deg] at (-0.25) should be [oblique 2.5deg]
PASS CSS Transitions: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0) should be [oblique 5deg]
PASS CSS Transitions: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0.3) should be [oblique 8deg]
PASS CSS Transitions: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0.6) should be [oblique 11deg]
PASS CSS Transitions: property <font-style> from [oblique 5deg] to [oblique 15deg] at (1) should be [oblique 15deg]
PASS CSS Transitions: property <font-style> from [oblique 5deg] to [oblique 15deg] at (1.5) should be [oblique 20deg]
FAIL CSS Transitions with transition: all: property <font-style> from [oblique 5deg] to [oblique 15deg] at (-2) should be [oblique -15deg] assert_equals: expected "oblique - 15deg " but got "normal "
PASS CSS Transitions with transition: all: property <font-style> from [oblique 5deg] to [oblique 15deg] at (-0.25) should be [oblique 2.5deg]
PASS CSS Transitions with transition: all: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0) should be [oblique 5deg]
PASS CSS Transitions with transition: all: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0.3) should be [oblique 8deg]
PASS CSS Transitions with transition: all: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0.6) should be [oblique 11deg]
PASS CSS Transitions with transition: all: property <font-style> from [oblique 5deg] to [oblique 15deg] at (1) should be [oblique 15deg]
PASS CSS Transitions with transition: all: property <font-style> from [oblique 5deg] to [oblique 15deg] at (1.5) should be [oblique 20deg]
FAIL CSS Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (-2) should be [oblique -15deg] assert_equals: expected "oblique - 15deg " but got "normal "
PASS CSS Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (-0.25) should be [oblique 2.5deg]
PASS CSS Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0) should be [oblique 5deg]
PASS CSS Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0.3) should be [oblique 8deg]
PASS CSS Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0.6) should be [oblique 11deg]
PASS CSS Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (1) should be [oblique 15deg]
PASS CSS Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (1.5) should be [oblique 20deg]
FAIL Web Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (-2) should be [oblique -15deg] assert_equals: expected "oblique - 15deg " but got "normal "
PASS Web Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (-0.25) should be [oblique 2.5deg]
PASS Web Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0) should be [oblique 5deg]
PASS Web Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0.3) should be [oblique 8deg]
PASS Web Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (0.6) should be [oblique 11deg]
PASS Web Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (1) should be [oblique 15deg]
PASS Web Animations: property <font-style> from [oblique 5deg] to [oblique 15deg] at (1.5) should be [oblique 20deg]
FAIL CSS Transitions: property <font-style> from [initial] to [inherit] at (-2) should be [oblique -40deg] assert_equals: expected "oblique - 40deg " but got "oblique 20deg "
FAIL CSS Transitions: property <font-style> from [initial] to [inherit] at (-0.25) should be [oblique -5deg] assert_equals: expected "oblique - 5deg " but got "oblique 20deg "
FAIL CSS Transitions: property <font-style> from [initial] to [inherit] at (0) should be [normal] assert_equals: expected "normal " but got "oblique 20deg "
FAIL CSS Transitions: property <font-style> from [initial] to [inherit] at (0.3) should be [oblique 6deg] assert_equals: expected "oblique 6deg " but got "oblique 20deg "
FAIL CSS Transitions: property <font-style> from [initial] to [inherit] at (0.6) should be [oblique 12deg] assert_equals: expected "oblique 12deg " but got "oblique 20deg "
PASS CSS Transitions: property <font-style> from [initial] to [inherit] at (1) should be [oblique 20deg]
FAIL CSS Transitions: property <font-style> from [initial] to [inherit] at (1.5) should be [oblique 30deg] assert_equals: expected "oblique 30deg " but got "oblique 20deg "
FAIL CSS Transitions with transition: all: property <font-style> from [initial] to [inherit] at (-2) should be [oblique -40deg] assert_equals: expected "oblique - 40deg " but got "oblique 20deg "
FAIL CSS Transitions with transition: all: property <font-style> from [initial] to [inherit] at (-0.25) should be [oblique -5deg] assert_equals: expected "oblique - 5deg " but got "oblique 20deg "
FAIL CSS Transitions with transition: all: property <font-style> from [initial] to [inherit] at (0) should be [normal] assert_equals: expected "normal " but got "oblique 20deg "
FAIL CSS Transitions with transition: all: property <font-style> from [initial] to [inherit] at (0.3) should be [oblique 6deg] assert_equals: expected "oblique 6deg " but got "oblique 20deg "
FAIL CSS Transitions with transition: all: property <font-style> from [initial] to [inherit] at (0.6) should be [oblique 12deg] assert_equals: expected "oblique 12deg " but got "oblique 20deg "
PASS CSS Transitions with transition: all: property <font-style> from [initial] to [inherit] at (1) should be [oblique 20deg]
FAIL CSS Transitions with transition: all: property <font-style> from [initial] to [inherit] at (1.5) should be [oblique 30deg] assert_equals: expected "oblique 30deg " but got "oblique 20deg "
FAIL CSS Animations: property <font-style> from [initial] to [inherit] at (-2) should be [oblique -40deg] assert_equals: expected "oblique - 40deg " but got "normal "
FAIL CSS Animations: property <font-style> from [initial] to [inherit] at (-0.25) should be [oblique -5deg] assert_equals: expected "oblique - 5deg " but got "normal "
PASS CSS Animations: property <font-style> from [initial] to [inherit] at (0) should be [normal]
FAIL CSS Animations: property <font-style> from [initial] to [inherit] at (0.3) should be [oblique 6deg] assert_equals: expected "oblique 6deg " but got "normal "
FAIL CSS Animations: property <font-style> from [initial] to [inherit] at (0.6) should be [oblique 12deg] assert_equals: expected "oblique 12deg " but got "oblique 20deg "
PASS CSS Animations: property <font-style> from [initial] to [inherit] at (1) should be [oblique 20deg]
FAIL CSS Animations: property <font-style> from [initial] to [inherit] at (1.5) should be [oblique 30deg] assert_equals: expected "oblique 30deg " but got "oblique 20deg "
FAIL Web Animations: property <font-style> from [initial] to [inherit] at (-2) should be [oblique -40deg] assert_equals: expected "oblique - 40deg " but got "normal "
FAIL Web Animations: property <font-style> from [initial] to [inherit] at (-0.25) should be [oblique -5deg] assert_equals: expected "oblique - 5deg " but got "normal "
PASS Web Animations: property <font-style> from [initial] to [inherit] at (0) should be [normal]
FAIL Web Animations: property <font-style> from [initial] to [inherit] at (0.3) should be [oblique 6deg] assert_equals: expected "oblique 6deg " but got "normal "
FAIL Web Animations: property <font-style> from [initial] to [inherit] at (0.6) should be [oblique 12deg] assert_equals: expected "oblique 12deg " but got "oblique 20deg "
PASS Web Animations: property <font-style> from [initial] to [inherit] at (1) should be [oblique 20deg]
FAIL Web Animations: property <font-style> from [initial] to [inherit] at (1.5) should be [oblique 30deg] assert_equals: expected "oblique 30deg " but got "oblique 20deg "
FAIL CSS Transitions: property <font-style> from [oblique 20deg] to [normal] at (-1) should be [oblique 40deg] assert_equals: expected "oblique 40deg " but got "normal "
FAIL CSS Transitions: property <font-style> from [oblique 20deg] to [normal] at (0) should be [oblique 20deg] assert_equals: expected "oblique 20deg " but got "normal "
FAIL CSS Transitions: property <font-style> from [oblique 20deg] to [normal] at (0.5) should be [oblique 10deg] assert_equals: expected "oblique 10deg " but got "normal "
PASS CSS Transitions: property <font-style> from [oblique 20deg] to [normal] at (1) should be [normal]
FAIL CSS Transitions: property <font-style> from [oblique 20deg] to [normal] at (1.5) should be [oblique -10deg] assert_equals: expected "oblique - 10deg " but got "normal "
FAIL CSS Transitions with transition: all: property <font-style> from [oblique 20deg] to [normal] at (-1) should be [oblique 40deg] assert_equals: expected "oblique 40deg " but got "normal "
FAIL CSS Transitions with transition: all: property <font-style> from [oblique 20deg] to [normal] at (0) should be [oblique 20deg] assert_equals: expected "oblique 20deg " but got "normal "
FAIL CSS Transitions with transition: all: property <font-style> from [oblique 20deg] to [normal] at (0.5) should be [oblique 10deg] assert_equals: expected "oblique 10deg " but got "normal "
PASS CSS Transitions with transition: all: property <font-style> from [oblique 20deg] to [normal] at (1) should be [normal]
FAIL CSS Transitions with transition: all: property <font-style> from [oblique 20deg] to [normal] at (1.5) should be [oblique -10deg] assert_equals: expected "oblique - 10deg " but got "normal "
FAIL CSS Animations: property <font-style> from [oblique 20deg] to [normal] at (-1) should be [oblique 40deg] assert_equals: expected "oblique 40deg " but got "oblique 20deg "
PASS CSS Animations: property <font-style> from [oblique 20deg] to [normal] at (0) should be [oblique 20deg]
FAIL CSS Animations: property <font-style> from [oblique 20deg] to [normal] at (0.5) should be [oblique 10deg] assert_equals: expected "oblique 10deg " but got "normal "
PASS CSS Animations: property <font-style> from [oblique 20deg] to [normal] at (1) should be [normal]
FAIL CSS Animations: property <font-style> from [oblique 20deg] to [normal] at (1.5) should be [oblique -10deg] assert_equals: expected "oblique - 10deg " but got "normal "
FAIL Web Animations: property <font-style> from [oblique 20deg] to [normal] at (-1) should be [oblique 40deg] assert_equals: expected "oblique 40deg " but got "oblique 20deg "
PASS Web Animations: property <font-style> from [oblique 20deg] to [normal] at (0) should be [oblique 20deg]
FAIL Web Animations: property <font-style> from [oblique 20deg] to [normal] at (0.5) should be [oblique 10deg] assert_equals: expected "oblique 10deg " but got "normal "
PASS Web Animations: property <font-style> from [oblique 20deg] to [normal] at (1) should be [normal]
FAIL Web Animations: property <font-style> from [oblique 20deg] to [normal] at (1.5) should be [oblique -10deg] assert_equals: expected "oblique - 10deg " but got "normal "
FAIL CSS Transitions: property <font-style> from [oblique -90deg] to [oblique 90deg] at (-2) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
FAIL CSS Transitions: property <font-style> from [oblique -90deg] to [oblique 90deg] at (-1) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
FAIL CSS Transitions: property <font-style> from [oblique -90deg] to [oblique 90deg] at (0) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
PASS CSS Transitions: property <font-style> from [oblique -90deg] to [oblique 90deg] at (0.5) should be [normal]
PASS CSS Transitions: property <font-style> from [oblique -90deg] to [oblique 90deg] at (1) should be [oblique 90deg]
FAIL CSS Transitions: property <font-style> from [oblique -90deg] to [oblique 90deg] at (1.5) should be [oblique 90deg] assert_equals: expected "oblique 90deg " but got "oblique 180deg "
FAIL CSS Transitions with transition: all: property <font-style> from [oblique -90deg] to [oblique 90deg] at (-2) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
FAIL CSS Transitions with transition: all: property <font-style> from [oblique -90deg] to [oblique 90deg] at (-1) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
FAIL CSS Transitions with transition: all: property <font-style> from [oblique -90deg] to [oblique 90deg] at (0) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
PASS CSS Transitions with transition: all: property <font-style> from [oblique -90deg] to [oblique 90deg] at (0.5) should be [normal]
PASS CSS Transitions with transition: all: property <font-style> from [oblique -90deg] to [oblique 90deg] at (1) should be [oblique 90deg]
FAIL CSS Transitions with transition: all: property <font-style> from [oblique -90deg] to [oblique 90deg] at (1.5) should be [oblique 90deg] assert_equals: expected "oblique 90deg " but got "oblique 180deg "
FAIL CSS Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (-2) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
FAIL CSS Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (-1) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
FAIL CSS Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (0) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
PASS CSS Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (0.5) should be [normal]
PASS CSS Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (1) should be [oblique 90deg]
FAIL CSS Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (1.5) should be [oblique 90deg] assert_equals: expected "oblique 90deg " but got "oblique 180deg "
FAIL Web Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (-2) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
FAIL Web Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (-1) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
FAIL Web Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (0) should be [oblique -90deg] assert_equals: expected "oblique - 90deg " but got "normal "
PASS Web Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (0.5) should be [normal]
PASS Web Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (1) should be [oblique 90deg]
FAIL Web Animations: property <font-style> from [oblique -90deg] to [oblique 90deg] at (1.5) should be [oblique 90deg] assert_equals: expected "oblique 90deg " but got "oblique 180deg "
FAIL An interpolation to inherit updates correctly on a parent style change. assert_equals: expected "oblique 10deg" but got "oblique 20deg"
